scrapers/banks/brac_bank.py
import requests
from bs4 import BeautifulSoup
from ..base_scraper import BaseScraper

class BracBankScraper(BaseScraper):
    BASE_URL = "https://www.bracbank.com"
    
    def get_bank_slug(self) -> str:
        return "brac-bank"
    
    def scrape_credit_cards(self) -> list[dict]:
        # Note: This is a placeholder structure. 
        # Real scraping requires inspecting the actual HTML structure of BRAC website.
        # For now, we return sample data to prove the pipeline works.
        
        cards = [
            {
                "name": "BRAC Bank Signature Card",
                "annual_fee": "BDT 10,000",
                "interest_rate": "20%",
                "features": ["Lounge Access", "Priority Pass"]
            },
            {
                "name": "BRAC Bank Platinum Card",
                "annual_fee": "BDT 5,000",
                "interest_rate": "20%",
                "features": ["BOGO Offers", "Reward Points"]
            }
        ]
        return cards
    
    def scrape_loan_products(self) -> list[dict]:
        return []
    
    def scrape_savings_rates(self) -> list[dict]:
        return []
